% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arcsinh_normalization.R
\name{Arcsinh_normalization}
\alias{Arcsinh_normalization}
\title{Transformation of raw intensity using arcsinh transform}
\usage{
Arcsinh_normalization(sce, cofactor_value = NULL)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{cofactor_value}{numeric value describing the arcsinh cofactor (by default set to 50)}
}
\value{
Returns a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object with a new assay slot called "Arcsinh_transformed_intensity"
}
\description{
This function transform the raw intensity
}
\examples{
sce = Arcsinh_normalization(sce,cofactor_value = 50)
}
