% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEdgeList.R
\name{buildEdgeList}
\alias{buildEdgeList}
\title{Create edgelist for spatial context graph}
\usage{
buildEdgeList(object, entry = NULL, img_id = NULL, combined = NULL)
}
\arguments{
\item{object}{a \code{SingleCellExperiment} or \code{SpatialExperiment}
object}

\item{entry}{single character specifying the \code{colData(object)} entry 
containing the \code{detectSpatialContext} output. If NULL, defaults to 
"spatial_context".}

\item{img_id}{single character specifying the \code{colData(object)} entry 
containing the unique image identifiers. If NULL, defaults to "sample_id".}

\item{combined}{should the edge-list be created on a cohort-level and not 
image-level? If NULL, defaults to TRUE.}
}
\value{
returns a data frame containing a symbolic edge list in the first
two columns ("from" and "to") and, if \code{combined = TRUE}, an additional 
column containing the \code{img_id} identifiers.
}
\description{
Function to create a symbolic edge list for spatial context 
graph construction. Based on single cell spatial context assignments, this 
function can operate on image- and cohort-level.
}
\examples{
TO DO

}
\author{
Lasse Meyer (\email{lasse.meyer@uzh.ch})
}
