% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSpatialGraph.R
\name{buildSpatialGraph}
\alias{buildSpatialGraph}
\title{Builds an interaction graph based on the cells' locations}
\usage{
buildSpatialGraph(
  object,
  img_id,
  type = c("expansion", "delauney", "knn"),
  k = NULL,
  threshold = NULL,
  coords = c("Pos_X", "Pos_Y"),
  name = NULL,
  directed = TRUE,
  BNPARAM = KmknnParam(),
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{object}{a \code{SingleCellExperiment} or \code{SpatialExperiment}
object}

\item{img_id}{single character indicating the \code{colData(object)} entry
containing the unique image identifiers.}

\item{type}{single character specifying the type of graph to be build.
Supported entries are \code{"expansion"} (default) to find interacting
cells via distance thresholding; \code{"delauney"} to find interactions via
delauney triangulation; \code{"knn"} to find the k nearest neighbouring
cells.}

\item{k}{(when \code{type = "knn"}) single numeric integer defining the
number of nearest neighbours to search for.}

\item{threshold}{(when \code{type = "expansion"}) single numeric specifying
the maximum distance for considering neighbours}

\item{coords}{character vector of length 2 specifying the names of the
\code{colData} (for a \code{SingleCellExperiment} object) or the
\code{spatialCoords} entries of the cells' x and y locations.}

\item{name}{single character specifying the name of the graph.}

\item{directed}{should the returned graph be directed? Only effects the k
nearest neighbour graph.}

\item{BNPARAM}{a \code{\link[BiocNeighbors]{BiocNeighborParam} object
defining the algorithm to use.}}

\item{BPPARAM}{a \code{\link[BiocParallel]{BiocParallelParam-class}} object
defining how to parallelise computations.}

\item{...}{additional parameters passed to the
\code{\link[BiocNeighbors]{findNeighbors}} function (\code{type =
"expansion"}), the \code{\link[RTriangle]{triangulate}} function
(\code{type = "delauney"}) or the \code{\link[BiocNeighbors]{findKNN}}
function (\code{type = "knn"})).}
}
\value{
returns a \code{SpatialExperiment} or \code{SingleCellExperiment}
containing the graph in form of a \code{SelfHits} object in
\code{colPair(object, name)}.
}
\description{
Function to define cell-cell interactions via distance-based
expansion, delauney triangulation or k nearest neighbour detection.
}
\section{Building an interaction graph}{


accessing by name
}

\section{Choosing the graph construction method}{


Default euclidean distance but manhattan and cosine supported via ...
}

\examples{
#TODO

}
\seealso{

}
\author{
Nils Eling (\email{nils.eling@dqbm.uzh.ch})
}
