% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSummary.R
\name{calculateSummary}
\alias{calculateSummary}
\title{Calculates marker intensity summary statistics given a cell-level factor}
\usage{
calculateSummary(x, split_by = NULL, exprs_values = "counts")
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{split_by}{character string or character list indicating which \code{colData(x)} entries to use for splitting.}

\item{exprs_values}{character string indicating from which \code{assays(x)} slot the intensity values can be extracted.
Default: exprs_values = "counts".}
}
\value{
Returns a \code{\link[SingleCellExperiment]{SingleCellExperiment}} that contains the following slots:
 \itemize{
   \item meanCounts stores the mean intensity values per factor level/ combination of levels
   \item medianCounts stores the median intensity values per factor level/ combination of levels
   \item varCounts stores the variance of intensity values per factor level/ combination of levels
 }
}
\description{
This function computes summary statistics (mean, median, and variance) within each factor level.
 When multiple entries are provided, summary statistics are calculated for eachcombination of factor entries.
}
\examples{
# TODO

}
\author{
Nils Eling \email{nils.eling@uzh.ch}
}
