% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_neighborhood_graph.R
\name{Compute_neighborhood_graph}
\alias{Compute_neighborhood_graph}
\title{Computation of a neighborhood graph from an SCE object}
\usage{
Compute_neighborhood_graph(sce, graph_type, graph_parameter = 10)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{graph_type}{type of graph to compute. Has to be chosen among the following values : "KNN","FRNN" or "Delaunay"}

\item{graph_parameter}{parameter for the graph computation. For the KNN it corresponds to the K parameter and for the FRNN to the radius value.}
}
\value{
Return an updated \code{\link[SingleCellExperiment]{SingleCellExperiment}} object with a list of graph, one per image
}
\description{
Computes a neighborhood graph using various strategies (KNN, FRNN, Delaunay triangulation)
}
\examples{
sce = Compute_neighborhood_graph(sce,graph_type = "KNN",graph_parameter = 10)


}
