% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_normalization.R
\name{Count_normalization}
\alias{Count_normalization}
\title{Normalisation and scaling of marker expression}
\usage{
Count_normalization(
  sce,
  perform_batch_correction = FALSE,
  batch_vector = NULL,
  residual_normalisation = "Anscombe"
)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{perform_batch_correction}{should covariates be regressed out ?}

\item{batch_vector}{vector describing the batch of each cell}

\item{residual_normalisation}{method for residual normalisation. Has to be chosen among "Anscombe","Pearson","Working" or "VST"}
}
\value{
Returns a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object with a new assay slot called "Normalized_intensity"
}
\description{
This function normalize and transform the intensity using a glm-based strategy
}
\examples{
sce = Count_normalization(sce,residual_normalisation = "Anscombe")
}
