% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_SCE.R
\name{Create_SCE}
\alias{Create_SCE}
\title{Generates an SCE data container for multiplexed single-cell imaging data}
\usage{
Create_SCE(List_data, dimension = "2D", Bit_mode = 16, N_core = 6)
}
\arguments{
\item{List_data}{a list object produced by Load_IMC_data}

\item{dimension}{the dimension of the data, either 2D or 3D}

\item{Bit_mode}{bit mode acquisition, usually set to 16}

\item{N_core}{number of cores used during parallelized computational steps}
}
\value{
returns an SCE object
}
\description{
This function takes the list produced by the Load_IMC_data() function and outputs an SCE object
}
\examples{
sce = Create_SCE(List_data,dimension = "2D",Bit_mode = 16,N_core = 6)

}
