% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_IMC_data.R
\name{Load_IMC_data}
\alias{Load_IMC_data}
\title{Load data generated by ImcSegmentationPipeline
  (https://github.com/BodenmillerGroup/ImcSegmentationPipeline)}
\usage{
Load_IMC_data(
  Path_to_cell_file,
  Path_to_panel_file,
  Name_target = "Target",
  Name_metal = "Metal.Tag",
  Regex_expression_channel = "Intensity_MeanIntensity_FullStackFiltered_c",
  DNA_channel = NULL,
  Use_for_clustering = NULL
)
}
\arguments{
\item{Path_to_cell_file}{path to the Cell.csv file produced by the ImcSegmentationPipeline}

\item{Path_to_panel_file}{path to the panel.csv file}

\item{Name_target}{name of the column from the panel.csv file that provides the gene name}

\item{Name_metal}{name of the column from the panel.csv file that provides the metal name}

\item{Regex_expression_channel}{full prefix of the cell.csv file column name that describe marker intensity}

\item{DNA_channel}{vector containing the name of the channels that measure cellular DNA amount}

\item{Use_for_clustering}{which channels will be used for cell clustering}
}
\value{
Returns a list object that can later be used to build a SCE object
}
\description{
This function loads data produced by the ImcSegmentationPipeline and output a list containing the different objects
}
\examples{
List_data = Load_IMC_data("Desktop/analysis/cpout/cell.csv",
                         Path_to_panel_file ="Desktop/analysis/cpout/panel.csv",
                         DNA_channel = c("Iridium191","Iridium193"),
                         Regex_expression_channel = "Intensity_MeanIntensity_FullStackFiltered_c",
                         Use_for_clustering = c("Myeloperoxidase.MPO","CD31","SMA","Cytokeratin.5",
                                                "Keratin.14","Vimentin","CD3","CD68", "Cytokeratin.8.18","CD45",
                                                "Carbonic.Anhydrase.IX","Fibronectin","Ki.67","CD20","CD44","pan.Cytokeratin"))

}
