% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_object_relationship.R
\name{Load_object_relationship}
\alias{Load_object_relationship}
\title{Loads an object relation csv file produced by the ImcSegmentationPipeline
  (https://github.com/BodenmillerGroup/ImcSegmentationPipeline)}
\usage{
Load_object_relationship(sce, Path_to_object_relationship)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{Path_to_object_relationship}{path to the Object relationships.csv file produced by the ImcSegmentationPipeline}
}
\value{
Return an updated \code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Loads an object relation csv file, constructs the corresponding graph and adds it to a previously defined sce object
}
\examples{
sce = Load_object_relationship(sce,Path_to_object_relationship = "Desktop/analysis/cpout/Object relationships.csv")


}
