% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_cluster_spatial.R
\name{Plot_cluster_spatial}
\alias{Plot_cluster_spatial}
\title{Visualization of cluster spatial distribution}
\usage{
Plot_cluster_spatial(sce, Image_number = 1)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}. Clustering of the cell should have been previously performed.}

\item{Image_number}{Number of name of the image/ROI to be plotted}
}
\value{
Return a plot
}
\description{
Plot the spatial distribution of a given gene normalised expression
}
\examples{
Plot_cluster_spatial(sce,Image_number = 1)


}
