% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Plot a low-dimensional embedding of single-cell data}
\usage{
plot_embedding(sce, selected_embedding = "UMAP", overlay_cluster = F)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{selected_embedding}{type of graph to compute. Has to be chosen among the following values : "KNN","FRNN" or "Delaunay"}

\item{overlay_cluster}{boolean value. If set to TRUE, will color the points according to their clustering}
}
\value{
Return a colored plot of a low-dimensional embedding
}
\description{
Plot a low-dimensional embedding of single-cell data that has been previously computed and is stored in a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\examples{
sce = plot_embedding(sce,selected_embedding = "UMAP",overlay_cluster=T)

}
