% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_gene_expression_spatial.R
\name{Plot_gene_expression_spatial}
\alias{Plot_gene_expression_spatial}
\title{Visualization of spatial gene expression}
\usage{
Plot_gene_expression_spatial(sce, Image_number = 1, Gene = NULL)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{Image_number}{Number of name of the image/ROI to be plotted}

\item{Gene}{Name of the gene to be plotted}
}
\value{
Return a plot
}
\description{
Plot the spatial distribution of a given gene normalised expression
}
\examples{
Plot_gene_expression_spatial(sce,Gene = "CD45")


}
