% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellCounts.R
\name{plotCellCounts}
\alias{plotCellCounts}
\title{Visualizes number of cells per factor level}
\usage{
plotCellCounts(x, colour_by = NULL, split_by = NULL, proportion = FALSE)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{colour_by}{character string corresponding to a \code{colData(x)} slot.
The resulting bar plot will be coloured by this factor.}

\item{split_by}{character string corresponding to a \code{colData(x)} slot.
Defines what will be plotted on the x-axis.}

\item{proportion}{logical if number of cells per \code{split_by} factor should be scaled to 1.
Default \code{FALSE}.}
}
\value{
Returns a \code{ggplot} barplot object that can be further modified following the \code{ggplot2} syntax.
}
\description{
This function plots the number of cells per chosen cell metadata factor.
}
\examples{
# TODO

}
\author{
Nils Eling \email{nils.eling@uzh.ch}
}
