% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.R
\name{plotDist}
\alias{plotDist}
\title{Visualizes distributions of marker intensities}
\usage{
plotDist(x, y = "rows", split_by = NULL, exprs_values = "counts", ...)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{y}{a character string indicating which factor to plot on the y-axis. Default: y = "rows", which plots a density for each feature.
Valid arguments are "rows" or any entry in the \code{colData(sce)} slot.}

\item{split_by}{character string corresponding to a \code{colData(x)} or \code{color_by = "rows"}, calling the rownames for \code{x}.
This feature will be used to facet wrap the plots.}

\item{exprs_values}{character string indicating from which \code{assays(x)} slot the intensity values can be extracted.
Default: exprs_values = "counts".}

\item{...}{further parameters for the \code{\link[ggridges]{ggridges}} function.}
}
\value{
Returns a \code{ggplot} object that can be further modified following the \code{ggplot2} syntax.
}
\description{
Ridge plot to visualize the distribution of cell intensities for each marker.
 These can be split an colored by cell-level metadata.
 By default, the intensity distributions for each marker will be displayed.
}
\examples{
# TODO

}
\author{
Nils Eling \email{nils.eling@uzh.ch}
}
