% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteractionsCozi.R
\name{plotInteractionsCozi}
\alias{plotInteractionsCozi}
\title{Plot interaction conditional ratio and z-score}
\usage{
plotInteractionsCozi(
  out,
  img_id = "group_by",
  zscore = "zscore",
  cond_ratio = "cond_ratio",
  sig = "sig",
  filter_sig = FALSE,
  zscore_lim = NULL,
  dot_size_lim = NULL
)
}
\arguments{
\item{out}{a data frame, usually the output from \code{testInteractions},
representing an edge list with columns \code{"group_by", "from_label",
"to_label", "zscore", "cond_ratio"}.}

\item{img_id}{a single character indicating the column in `out` that
represents the image or sample ID.}

\item{zscore}{single character indicating the name of the column in `out`
that contains the z-score values. Defaults to "zscore".}

\item{cond_ratio}{single character indicating the name of the column in
`out` that contains the conditional ratio values. Defaults to "cond_ratio".}

\item{filter_sig}{logical indicating if results should be filtered based on significance testing. 
If TRUE, only interactions with sig == TRUE will be plotted. Defaults to FALSE.}

\item{zscore_lim}{a numeric vector of length 2 specifying the limits for the
z-score color scale.}

\item{dot_size_lim}{a numeric vector of length 2 specifying the limits for the
dot size scale.}
}
\value{
returns a \code{ggplot} object.
}
\description{
Function to plot a dot plot visualizing the conditional ratio and
z-score of cell-cell interactions calculated with testInteractinos() with method= "conditional"
}
\examples{
set.seed(22)
library(cytomapper)
library(BiocParallel)
library(data.table)
data(pancreasSCE)

## 1. countInteractions or testInteractions with setting method = "conditional"
sce  <- buildSpatialGraph(pancreasSCE, img_id = "ImageNb", type = "knn", k = 3)

count_out <- countInteractions(sce,
                               group_by = "ImageNb",
                               label = "CellType",
                               method = "conditional", 
                               colPairName = "knn_interaction_graph")

test_out <- testInteractions(sce, 
                             group_by = "ImageNb",
                             label = "CellType", 
                             method = "conditional", 
                             colPairName = "knn_interaction_graph", 
                             iter = 100, 
                             p_threshold = 0.5, 
                             BPPARAM = SerialParam(RNGseed = 123))

## 2. Plot interactions as dotplot

# default                
plotInteractionsCozi(test_out)

# adjust zscore and dot size limits
plotInteractionsCozi(test_out,
                    zscore_lim = c(-50, 50),
                   dot_size_lim = c(0, 0.5))  

# filter for significant interactions only
plotInteractionsCozi(test_out,
                   filter_sig = TRUE)   
                  
}
\seealso{
\code{\link{countInteractions}} for counting (but not testing) cell-cell
interactions per grouping level.
\code{\link{testInteractions}} for testing cell-cell 
interactions per grouping level.
}
\author{
Chiara Schiller (\email{chiara.schiller@uni-heidelberg.de})
}
