% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialContext.R
\name{plotSpatialContext}
\alias{plotSpatialContext}
\title{Plot spatial context graph}
\usage{
plotSpatialContext(
  edges,
  object,
  entry = NULL,
  img_id = NULL,
  combined = NULL,
  node_color_by = NULL,
  node_size_by = NULL,
  node_color_fix = NULL,
  node_size_fix = NULL,
  node_label_repel = TRUE,
  node_label_color_by = NULL,
  node_label_color_fix = NULL,
  draw_edges = TRUE,
  edge_color_fix = NULL
)
}
\arguments{
\item{edges}{data frame containing a symbolic edge list in the first two 
columns.}

\item{object}{a \code{SingleCellExperiment} or \code{SpatialExperiment}
object}

\item{entry}{single character specifying the \code{colData(object)} entry 
containing the \code{detectSpatialContext} output. If NULL, defaults to 
"spatial_context".}

\item{img_id}{single character specifying the \code{colData(object)} entry 
containing the unique image identifiers. If NULL, defaults to "sample_id".}

\item{combined}{Is the provided edge list created on cohort-level (two 
columns) and not image-level (three columns)? If NULL, defaults to TRUE.}

\item{node_color_by}{single character from \code{c("name","Freq","n_samples")} 
by which the nodes should be colored.}

\item{node_size_by}{single character from \code{c("Freq","n_samples")} 
by which the size of the nodes are defined.}

\item{node_color_fix}{single character specifying the color of all nodes.}

\item{node_size_fix}{single numeric specifying the size of all nodes.}

\item{node_label_repel}{should nodes be labelled? Defaults to TRUE.}

\item{node_label_color_by}{single character from \code{c("name","Freq","n_samples")} 
by which the node labels should be colored.}

\item{node_label_color_fix}{single character specifying the color of all node 
labels.}

\item{draw_edges}{should edges be drawn between nodes? Defaults to TRUE.}

\item{edge_color_fix}{single character specifying the color of all edges.}
}
\value{
returns a \code{ggplot} object.
}
\description{
Function to plot directed spatial context graphs based on symbolic
edge-lists, which operates on image- and cohort-level. User can specify node, 
node_label and edge aesthetics.
}
\examples{
TO DO

}
\author{
Lasse Meyer (\email{lasse.meyer@uzh.ch})
}
