% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSCEfromTXT.R
\name{prepareSCEfromTXT}
\alias{prepareSCEfromTXT}
\title{Generates a SingleCellExperiment from .txt files}
\usage{
prepareSCEfromTXT(
  txt_list,
  metadata_cols = c("Start_push", "End_push", "Pushes_duration", "X", "Y", "Z"),
  verbose = TRUE
)
}
\arguments{
\item{txt_list}{a named list containing the read-in txt files for each spot. 
Entries to \code{txt_list} need to be coercible to \code{data.frame} objects.}

\item{metadata_cols}{character vector indicating which column entries of the
.txt files should be saved in the \code{colData(sce)} slot.}

\item{verbose}{logical indicating if additional information regarding the
spotted and acquired masses should be shown.}
}
\value{
returns a SCE object
}
\description{
Helper function to process raw .txt files acquired by the
  Hyperion system into a \code{\linkS4class{SingleCellExperiment}} container.
  This function is mainly used to read-in data generated from a "spillover
  slide". Here, each .txt file contains the measurements of multiple pixels
  for a single stain across all open channels.
}
\examples{

}
