% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeNeighbors.R
\name{summarizeNeighbors}
\alias{summarizeNeighbors}
\title{Function to summarize the neighborhood from each cell.}
\usage{
summarizeNeighbors(
  object,
  colPairName,
  summarize_by = c("celltypes", "expression"),
  group = NULL,
  assay_type = NULL,
  subset_row = NULL,
  summaryStats = NULL,
  name = NULL
)
}
\arguments{
\item{object}{a \code{SingleCellExperiment} object}

\item{colPairName}{single character indicating the \code{colPair(object)} entry
containing the the neighbor information.}

\item{summarize_by}{character specifying whether the neighborhood should be summarized by cellular
features from the \code{colData(object)} or by the marker expression of the neighboring cells.}

\item{group}{if selecting "celltypes" in \code{summarize_by} a single character specifying the \code{colData(object)} entry containing the
cellular features that should be summarized in the neighborhood.
Supported entries are of type character.}

\item{subset_row}{if selecting "expression" in \code{summarize_by} a character vector specifying the entries from \code{rownames(object)}
to use for the summary statistics.}

\item{summaryStats}{if selecting "expression" in \code{summarize_by} then a single character specifying the summary statistics should be provided.
Supported entries are "mean", "median", "sd", "var".}

\item{name}{single character specifying the name of the data frame to be saved in the \code{colData(object)}.}
}
\value{
returns a \code{SingleCellExperiment}
containing the data frame in form of a \code{DataFrame} object in
\code{colData(object)}.
}
\description{
Function to generates a data frame that either summarized the  count of cells of a specified type
in the selected neighborhood of each cell or summarizes the marker expression of all cells in the selected neighborhood of each cell.
}
\author{
Daniel Schulz (\email{daniel.schulz@uzh.ch})
}
