% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAP_embedding.R
\name{UMAP_embedding}
\alias{UMAP_embedding}
\title{Computation of a UMAP embedding}
\usage{
UMAP_embedding(sce, metric = "correlation", n_neighbors = 30)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{path}{to the Object relationships.csv file produced by the ImcSegmentationPipeline}
}
\value{
Return an updated \code{\link[SingleCellExperiment]{SingleCellExperiment}} object with a list of graph, one per image
}
\description{
Computes a UMAP embedding
}
\examples{
sce = Compute_neighborhood_graph(sce,graph_type = "KNN",graph_parameter = 10)


}
