% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSpatialContext.R
\name{detectSpatialContext}
\alias{detectSpatialContext}
\title{Detect the spatial context of each cell based on its neighborhood}
\usage{
detectSpatialContext(object, entry = NULL, threshold = 0.9, name = NULL)
}
\arguments{
\item{object}{a \code{SingleCellExperiment} or \code{SpatialExperiment}
object}

\item{entry}{single character specifying the \code{colData(object)}
entry containing the \code{aggregateNeighbors} DataFrame output. If NULL, 
defaults to "aggregatedNeighbors".}

\item{threshold}{single numeric between 0 and 1 that specifies the fraction 
threshold for SC assignment. Defaults to 0.9.}

\item{name}{single character specifying the name of the output saved in 
\code{colData(object)}.}
}
\value{
returns an object of \code{class(object)} containing a new column 
entry to \code{colData(object)[[name]]}
}
\description{
Function to detect the spatial context (SC) of each cell. 
Based on its sorted (high-to-low) cellular neighborhood (CN) fractions in a 
k-nearest neighbor graph, the SC of each cell is assigned as the set of CNs 
that cumulatively exceed a user-defined fraction threshold. 

The term was coined by Bhates et al. (Cell Systems, 2022) 
<https://doi.org/10.1016/j.cels.2021.09.012> and describes tissue regions 
in which distinct CNs may be interacting.
}
\examples{
TO DO

}
\author{
Lasse Meyer (\email{lasse.meyer@uzh.ch})
}
