% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodPermTest.R
\name{neighborhoodPermTest}
\alias{neighborhoodPermTest}
\title{Performs perturbations to test if two cell-types interact more or less frequently than random.}
\usage{
neighborhoodPermTest(object, method = c("histocat", "classic"))
}
\arguments{
\item{object}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{method}{which aggregation method to use}
}
\description{
This function computes summary statistics (mean, median, and variance) within each factor level.
 When multiple entries are provided, summary statistics are calculated for eachcombination of factor entries.
}
\examples{
# TODO

}
\author{
Vito Zantonelli, adapted by Nils Eling \email{nils.eling@uzh.ch}
}
