% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchDetection.R
\name{patchDetection}
\alias{patchDetection}
\title{Function to detect patches containing defined cell types}
\usage{
patchDetection(
  object,
  patch_cells,
  colPairName,
  min_patch_size = 1,
  name = "patch_id",
  expand_by = 0,
  coords = c("Pos_X", "Pos_Y"),
  convex = FALSE,
  img_id = NULL,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{object}{a \code{SingleCellExperiment} or \code{SpatialExperiment}
object}

\item{patch_cells}{logical vector of length equal to the number of cells
contained in \code{object}. \code{TRUE} entries define the cells to consider
for patch detection (see Details).}

\item{colPairName}{single character indicating the \code{colPair(object)}
entry containing the neighbor information.}

\item{min_patch_size}{single integer indicating the minimum number of 
connected cells that make up a patch before expansion.}

\item{name}{single character specifying the \code{colData} entry storing
the patch IDs in the returned object.}

\item{expand_by}{single numeric indicating in which vicinity range cells
should be considered as belonging to the patch (see Details).}

\item{coords}{character vector of length 2 specifying the names of the
\code{colData} (for a \code{SingleCellExperiment} object) or the
\code{spatialCoords} entries of the cells' x and y locations.}

\item{convex}{should the convex hull be computed before expansion? Default:
the concave hull is computed.}

\item{img_id}{single character indicating the \code{colData(object)} entry
containing the unique image identifiers.}

\item{BPPARAM}{a \code{\link[BiocParallel]{BiocParallelParam-class}} object
defining how to parallelize computations.}
}
\value{
An object of \code{class(object)} containing a patch ID for each 
cell in \code{colData(object)[[name]]}.
}
\description{
Function to detect spatial clusters of defined types of cells.
By defining a certain distance threshold, all cells within the vicinity
of these clusters are detected as well.
}
\section{Detecting patches of defined cell types}{

Explain NAs
}

\examples{
library(cytomapper)
data(pancreasSCE)

pancreasSCE <- buildSpatialGraph(pancreasSCE, img_id = "ImageNb", 
                                 type = "expansion", threshold = 20)

pancreasSCE <- patchDetection(pancreasSCE, 
                              patch_cells = pancreasSCE$CellType == "celltype_A",
                              colPairName = "knn_interaction_graph")

  
}
\references{
\href{https://www.biorxiv.org/content/10.1101/2021.07.29.454093v1}{
Hoch, T. et al., Multiplexed Imaging Mass Cytometry of Chemokine Milieus in 
Metastatic Melanoma Characterizes Features of Response to Immunotherapy., 
bioRxiv 2021}
}
\author{
Tobias Hoch

adapted by Nils Eling (\email{nils.eling@dqbm.uzh.ch})
}
