% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatial.R
\name{plotSpatial}
\alias{plotSpatial}
\title{Visualizes the spatial locations of cells}
\usage{
plotSpatial(
  object,
  img_id,
  coords = c("Pos_X", "Pos_Y"),
  node_color_by = NULL,
  node_shape_by = NULL,
  node_size_by = NULL,
  edge_color_by = NULL,
  edge_size_by = NULL,
  draw_edges = FALSE,
  arrow = NULL,
  colPairName = NULL,
  ncols = NULL,
  nrows = NULL
)
}
\arguments{
\item{object}{a \code{SingleCellExperiment} or \code{SpatialExperiment}
object.}

\item{img_id}{single character indicating the \code{colData(object)} entry
containing the unique image identifiers.}

\item{coords}{character vector of length 2 specifying the names of the
\code{colData} (for a \code{SingleCellExperiment} object) or the
\code{spatialCoords} entries of the cells' x and y locations.}

\item{node_color_by}{single character indicating the \code{colData(object)}
entry by which the nodes (cell locations) should be colored.}

\item{node_shape_by}{single character indicating the \code{colData(object)}
entry by which the shape of the nodes are defined.}

\item{node_size_by}{single character indicating the \code{colData(object)}
entry by which the size of the nodes are defined.}

\item{edge_color_by}{single character indicating by which to color the edges.
See details for more information.}

\item{edge_size_by}{single character determining the size of the edges.
See details for more information.}

\item{draw_edges}{should cell-cell interactions be drawn as edges between
nodes?}

\item{arrow}{a \code{\link[grid]{arrow}} object specifying how to draw arrows
between cells.}

\item{colPairName}{single character specifying from which 
\code{colPair(object)} slot to retrieve the cell-cell pairings.}

\item{ncols}{number of columns of the grid to arrange individual images.}

\item{nrows}{number of rows of the grid to arrange individual images.}
}
\value{
returns a \code{ggplot} object.
}
\description{
A general function to plot spatial locations of cells while 
specifying color, shape, size. Cell-cell interactions can be visualized
in form of edges between points.
}
\section{Visualizing cell locations and cell-cell interactions}{
 
By default, the cells' locations are visualized in form of points (here also
referred to as "nodes") on a 2-dimensional plane. The cells' coordinates are
extracted either from \code{colData(object)} slot (for a
\code{SingleCellExperiment} input object) or from the
\code{spatialCoords(object)} slot (for a \code{SpatialExperiment} input
object). Node aesthetics are controlled by setting \code{node_color_by}, 
\code{node_shape_by} and \code{node_size_by}. 

When \code{draw_edges = TRUE}, cell-cell interactions are visualized in form
of edges between nodes. For this, \code{object} needs to contain 
column pairings in \code{colPair(object, colPairName)}. Edge color and size
can be set by specifying either an entry in 
\code{mcols(colPair(object, colPairName))} (edge attributes) or in 
\code{colData(object)}. In the latter case, edges are colored by attributes 
associated to the "from" node. 

Arrows for displaying directed graphs can be drawn by supplying a 
\code{\link[grid]{arrow}} object. Arrow attributes can be set within this 
class.
}

\examples{
library(cytomapper)
data(pancreasSCE)

sce <- buildSpatialGraph(pancreasSCE, img_id = "ImageNb", 
                         type = "knn", k = 3)

plotSpatial(sce, 
            node_color_by = "CellType",
            node_shape_by = "ImageNb",
            node_size_by = "Area")
  
}
\seealso{

}
\author{
Nils Eling (\email{nils.eling@dqbm.uzh.ch})
}
