% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMedianHeatmap.R
\name{plotSpotHeatmap}
\alias{plotSpotHeatmap}
\title{Visualizes the median pixel intensities per spot and channel}
\usage{
plotSpotHeatmap(
  object,
  spot_id = "sample_id",
  channel_id = "channel_name",
  assay_type = "counts",
  statistic = "median",
  log = TRUE,
  threshold = NULL,
  order_metals = TRUE,
  ...
)
}
\arguments{
\item{object}{a \code{SingleCellExperiment} object containing pixel
intensities per channel. Individual pixels are stored as columns and
channels are stored as rows.}

\item{spot_id}{character string indicating which \code{colData(object)} entry
stores the isotope names of the spotted metal. Entries should be of the 
form (mt)(mass) (e.g. Sm152 for Samarium isotope with the atomic mass 152).}

\item{channel_id}{character string indicating which \code{rowData(object)} 
entry contains the isotope names of the acquired channels.}

\item{assay_type}{character string indicating which assay to use.}

\item{log}{should the aggregated pixel intensities be \code{log10(x + 1)} 
transformed?}

\item{order_metals}{should the metals be ordered based on spotted mass?}

\item{...}{arguments passed to \code{pheatmap}.}

\item{statistics}{the statistic to use when aggregating channels per spot}

\item{theshold}{single numeric indicating the aggregated pixel threshold.
This facilitates the identification of spots with low aggregated intensities.}
}
\value{
returns a SCE object
}
\description{
Helper function for estimating the spillover matrix. This
function visualizes the median pixel intensities per spot (columns) and per
channel (rows).
}
\examples{

}
