% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cpout.R
\name{read_cpout}
\alias{read_cpout}
\title{Generates an SCE data container for single-cell IMC data generated by
  ImcSegmentationPipeline
  (https://github.com/BodenmillerGroup/ImcSegmentationPipeline)}
\usage{
read_cpout(
  cells = NULL,
  cell_meta = c("ImageNumber", "ObjectNumber"),
  image_meta = NULL,
  panel_meta = NULL,
  metal_tag = "Metal.Tag",
  channel_number = "channel",
  target_names = NULL,
  imageNumber = "ImageNumber",
  objectNumber = "ObjectNumber",
  measurement_column = NULL,
  scaling_column = NULL,
  metadata = NULL
)
}
\arguments{
\item{cells}{\code{data.frame} containing single-cell measurements (cell.csv)}

\item{cell_meta}{character vector containing column names of relevant meta
data stored in `cells`, must at least include a sort of `CellNumber` and
`ImageNumber`}

\item{image_meta}{\code{data.frame} containing meta data for the acquired
samples (image.csv)}

\item{panel_meta}{\code{data.frame} containing meta data for the applied
panel}

\item{metal_tag}{\code{colnames} of metal tag column in `panel_meta` (default
= "Metal.Tag")}

\item{channel_number}{\code{colnames} of column that contains the channel number in `panel_meta` (default = "channel")}

\item{target_names}{\code{colnames} of column that contains the antibody/target name in `panel_meta`}

\item{imageNumber}{name of column containing an sample identifier present in
all provided input}

\item{objectNumber}{name of column containing an object number for individual
cells}

\item{measurement_column}{shared \code{colnames} pattern of \code{cells}
columns storing channel measurement (e.g.
Intensity_MeanIntensityCorrected_FullStackFiltered)}

\item{scaling_column}{\code{colnames} of scaling column storing the applied
scaling factor (e.g. Scaling_FullStack)}

\item{metadata}{(optional) \code{data.frame} containing further metadata
associated to cells, must include `imageID`}
}
\value{
returns SCE object
}
\description{
This function takes various inputs and stores it in an
  \code{\link[SingleCellExperiment]{SingleCellExperiment}} data container
}
\examples{
#sce <- generateSCE(cells = cells,
#       cell_meta = c("ImageNumber", "ObjectNumber"),
#       image_meta = image_mat,
#       panel_meta = panel_mat,
#       metal_tag = "Metal.Tag",
#       channel_number = "channel", 
#       target_names = "clean_target",
#       imageNumber = "ImageNumber",
#       objectNumber = "ObjectNumber",
#       measurement_column = "Intensity_MeanIntensityComp_FullStackFiltered",
#       scaling_column = "Scaling_FullStack")

}
\author{
Tobias Hoch

Nils Eling (\email{nils.eling@dqbm.uzh.ch})
}
