% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImagefromTXT.R
\name{readImagefromTXT}
\alias{readImagefromTXT}
\title{Reads one or multiple .txt files into a CytoImageList object}
\usage{
readImagefromTXT(
  path,
  pattern = ".txt$",
  channel_pattern = "[A-Za-z]{1,2}[0-9]{2,3}Di",
  index_names = c("X", "Y", "Z"),
  BPPARAM = SerialParam()
)
}
\arguments{
\item{path}{Full path to where the individual .txt files are located. This is
usualy the path where the .mcd file is located.}

\item{pattern}{pattern to select which files should be read in (default
\code{".txt$"}).}

\item{channel_pattern}{regular expression to select the channel names from
the files.}

\item{BPPARAM}{parameters for parallelised reading in of images. 
This is only recommended for very large images.}
}
\value{
returns a \code{\linkS4class{CytoImageList}} object containing one
\code{\linkS4class{Image}} object per .txt file.
}
\description{
Reader function to generate \code{\linkS4class{Image}} objects
in form of a \code{\linkS4class{CytoImageList}} container from .txt files.
}
\section{Imaging mass cytometry .txt files}{


Explain what they are
}

\examples{

}
\seealso{
\code{\linkS4class{CytoImageList}} for the container
\code{\link[BiocParallel]{MulticoreParam}} for parallelised processing
\code{\linkS4class{Image}} for the multi-channel image object
\code{vignette("cytomapper")} for visualization of multi-channel images
}
\author{
Nils Eling (\email{nils.eling@dqbm.uzh.ch})
}
