% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeEnvironment.R
\name{summarizeNeighbors}
\alias{summarizeNeighbors}
\title{Generates a data frame with the count of cells of a specified type (e.g. clustered celltypes) in the selected neighborhood of each cell.}
\usage{
summarizeNeighbors(object, colPairName, group, name = NULL)
}
\arguments{
\item{object}{a \code{SingleCellExperiment} object}

\item{colPairName}{single character indicating the \code{colPair(object)} entry
containing the the neighbor information.}

\item{group}{single character specifying the \code{colData(object)} entry containing the
cellular features that should be summarized in the neighborhood.
Supported entries are of type character.}

\item{name}{single character specifying the name of the data frame.}
}
\value{
returns a \code{SingleCellExperiment}
containing the data frame in form of a \code{DataFrame} object in
\code{colData(object)}.
}
\description{
Function to generates a data frame with the count of cells of a specified type (e.g. clustered celltypes)
in the selected neighborhood of each cell.
}
\author{
Daniel Schulz (\email{daniel.schulz@uzh.ch})
}
